+++ crates/store/src/genesis/config/errors.rs
@@ -1,13 +1,7 @@
 use std::path::PathBuf;
 
 use miden_protocol::account::AccountId;
-use miden_protocol::errors::{
-    AccountDeltaError,
-    AccountError,
-    AssetError,
-    FeeError,
-    TokenSymbolError,
-};
+use miden_protocol::errors::{AccountDeltaError, AccountError, AssetError, TokenSymbolError};
 use miden_protocol::utils::serde::DeserializationError;
 use miden_standards::account::faucets::FungibleFaucetError;
 use miden_standards::account::policies::TokenPolicyManagerError;
@@ -65,8 +59,6 @@ pub enum GenesisConfigError {
     IssuanceOverflow,
     #[error("missing fee faucet for native asset {0}")]
     MissingFeeFaucet(TokenSymbolStr),
-    #[error("fee error")]
-    FeeError(#[from] FeeError),
     #[error("faucet account of {0} is not a fungible faucet")]
     NativeAssetFaucetIsNotPublic(TokenSymbolStr),
     #[error("faucet account of {0} is not public")]